Added a preclassical calculator [skip hazardlib]
