+++ openquake/calculators/classical.py
@@ -18,13 +18,14 @@
 
 from __future__ import division
 import math
+import time
 import logging
 import operator
 import numpy
 
 from openquake.baselib import parallel, hdf5
 from openquake.baselib.python3compat import encode
-from openquake.baselib.general import AccumDict, block_splitter
+from openquake.baselib.general import AccumDict, block_splitter, groupby
 from openquake.hazardlib.calc.hazard_curve import classical, ProbabilityMap
 from openquake.hazardlib.stats import compute_pmap_stats
 from openquake.hazardlib import source
@@ -236,6 +237,38 @@ class PSHACalculator(base.HazardCalculator):
                     # only for the case of a single realization
                     self.datastore['disagg_by_src/source_id'] = numpy.array(
                         sorted(data), grp_source_dt)
+
+
+def count_ruptures(sources, srcfilter, gsims, param, monitor):
+    """
+    Count the number of ruptures contained in the given sources by applying a
+    raw source filtering on the integration distance. Return a dictionary
+    src_group_id -> {}.
+    All sources must belong to the same tectonic region type.
+    """
+    dic = groupby(sources, lambda src: src.src_group_ids[0])
+    acc = AccumDict({grp_id: {} for grp_id in dic})
+    acc.eff_ruptures = {grp_id: 0 for grp_id in dic}
+    acc.calc_times = AccumDict(accum=numpy.zeros(4))
+    for grp_id in dic:
+        for src in sources:
+            t0 = time.time()
+            src_id = src.source_id.split(':')[0]
+            sites = srcfilter.get_close_sites(src)
+            if sites is not None:
+                acc.eff_ruptures[grp_id] += src.num_ruptures
+                dt = time.time() - t0
+                acc.calc_times[src_id] += numpy.array(
+                    [src.weight, len(sites), dt, 1])
+    return acc
+
+
+@base.calculators.add('preclassical')
+class PreClassicalCalculator(PSHACalculator):
+    """
+    Classical PSHA calculator
+    """
+    core_task = count_ruptures
 
 
 def fix_ones(pmap):
